Removed GNUmake-related files from source control Removed 3rd-party JARs from source control Re-introduced original BouncyCastle-dependent PKCS5KeyFile class Re-named non-BouncyCastle-dependent PKCS8KeyFile class to PKCS5KeyFile, since it really only supports PKCS5 formats anyway
