--- GNUmakefile
@@ -1,46 +1,0 @@
-Default: all
-
-TOP=$(realpath .)
-include $(TOP)/../DeveloperTools/install/common.mk
-
-RSRC=rsrc
-LIBDIR=$(RSRC)/lib
-LIB=$(subst $(SPACE),$(CLN),$(filter %.jar %.zip, $(wildcard $(LIBDIR)/*)))
-BUILD=build
-SRC=src/main/java
-DOCS=docs
-CLASSPATH="$(CLASSLIB)$(CLN)$(LIB)$(CLN)$(SRC)"
-CWD=$(shell pwd)
-
-include classes.mk
-
-CLASS_FILES:=$(foreach class, $(CLASSES), $(BUILD)/$(subst .,/,$(class)).class)
-PACKAGES=$(sort $(basename $(CLASSES)))
-PACKAGEDIRS=$(subst .,/,$(PACKAGES))
-
-all: sshj.jar
-
-sshj.jar: classes
-	$(JAR) cvf $@ -C $(BUILD)/ .
-
-javadocs:
-	mkdir -p $(DOCS)
-	$(JAVA_HOME)/bin/javadoc -d $(DOCS) -classpath $(CLASSPATH) $(PACKAGES)
-
-clean:
-	rm -rf $(BUILD)
-
-classes: classdirs $(CLASS_FILES)
-
-install: all
-	cp sshj.jar $(TOP)/../jOVAL-Commercial/components/wsmv/winrs/rsrc/lib
-	cp sshj.jar $(TOP)/../jOVAL-Commercial/components/provider/remote/rsrc/lib
-	cp sshj.jar $(TOP)/../jOVAL-Commercial/components/sdk/dist/3rd-party
-
-classdirs: $(foreach pkg, $(PACKAGEDIRS), $(BUILD)/$(pkg)/)
-
-$(BUILD)/%.class: $(SRC)/%.java
-	$(JAVAC) $(JAVACFLAGS) -d $(BUILD) -classpath $(CLASSPATH) $<
-
-$(BUILD)/%/:
-	mkdir -p $(subst PKG,,$@)
